made sure that the input system cleans itself up properly.
