+++ Assets/MixedRealityToolkit/_Core/Interfaces/InputSystem/IPointerResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a778d0f4f5894683a842581d81a4492d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8ce0387fe0b4c824aa16dee6d56afd96, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
