+++ Assets/MixedRealityToolkit/_Core/Interfaces/InputSystem/ITeleportTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1448e888ea240bcb71795c221952904
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8ce0387fe0b4c824aa16dee6d56afd96, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
